SO_REUSEPORT needs to be applied in PH_PREBIND stage.
OPT_SO_REUSEPORT added in Linux 3.9 (as per socket(7)), needs to be
applied in PREBIND stage, otherwise you get this error

```
2015/05/27 18:53:02 socat[9032] E bind(5, {AF=2 0.0.0.0:4444}, 16): Address already in use
```

This is not supposed to happen since SO_REUSEPORT implies multiplexing
on same (address,port).

Hence, making so-reuseport to be PH_PREBIND. This way it is applied to
socket before the bind and it works as intended.

Tested as in https://gist.github.com/ronin13/e4f00da7c11e897383aa